+++ Assets/Prefabs/Input/Player.prefab
@@ -675,7 +675,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1245027742680820363}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -3.48, y: 5.67, z: -2.14}
+  m_LocalPosition: {x: -2.02, y: 5.67, z: 2.96}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1367,7 +1367,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5583479561026228546}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 3.5, y: 5.81, z: -0.86}
+  m_LocalPosition: {x: 3.26, y: 4.32, z: 0.48}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -1514,7 +1514,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8105111567943840941}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 1.809, z: -0.062}
+  m_LocalPosition: {x: 0, y: 1.837, z: -0.265}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -3067,15 +3067,16 @@ MonoBehaviour:
   leftArmRoot: {fileID: 7410943623118469505}
   leftArmRootOffset: -70
   leftArmElbow: {fileID: 1755913342065706940}
-  leftArmElbowOffset: -1.6
+  leftArmElbowOffset: -12.6
   leftArmHand: {fileID: 3680236445534610840}
   leftHandIKPole: {fileID: 3466674245403287988}
   rightArmRoot: {fileID: 332192596224579183}
-  rightArmRootOffset: 70
+  rightArmRootOffset: 180
   rightArmElbow: {fileID: 8174527586192360440}
-  rightArmElbowOffset: 1.6
+  rightArmElbowOffset: 12.6
   rightArmHand: {fileID: 4604144610022740314}
   rightHandIKPole: {fileID: 6133131847345276541}
+  modelScaleOffset: 0.04
 --- !u!1 &665535824471600874 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 2638215920088325915, guid: 27ac4408b426ae745a1792dde18c79ba, type: 3}
