+++ Assets/Scripts/Utils/GenericTwoBodyIK.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3fb3d10cf598cab47b47bfe85d92a6a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
